Revert "move relay notification code"
This reverts commit 466a15904cd7d9b73f43c71cfeabaa1f37adb466.
